+++ src/main/kotlin/MultiplesOf3And5.kt
@@ -7,10 +7,10 @@ class MultiplesOf3And5: Solution {
         const val ROOF = 999
     }
 
-    override val rightSolution: Int = 233168
+    override val rightSolution: Long = 233168
 
-    override fun solve(): Int {
-        var sum = 0
+    override fun solve(): Long {
+        var sum = 0L
         for (i in FLOOR..ROOF){
             if(i % 3 == 0 || i % 5 == 0){
                 sum += i
