+++ src/main/kotlin/TenThousandAndOnePrime.kt
@@ -5,8 +5,8 @@ class TenThousandAndOnePrime: Solution {
 		const val GOAL = 10001
 	}
 
-	override val rightSolution = 104743
-	override fun solve(): Int{
+	override val rightSolution = 104743L
+	override fun solve(): Long{
 
 		var position = 6 // in problem description it says that the 6th prime is 13, so I can move on from here
 		var current = 14L
@@ -18,6 +18,6 @@ class TenThousandAndOnePrime: Solution {
 			if(position != GOAL) current++
 		}
 
-		return current.toInt()
+		return current
 	}
  }
